ai_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
chromadb
pypdf
openai
